requirements.txt
git+https://github.com/edx/xblock-utils.git@v1.0.2#egg=xblock-utils==1.0.2
libsass==0.10.0
-e .
